CGIV-7089: updated all validation rules for non-required boolean fields to also allow them to be empty.
This follows on from the previous change to lock zend-inputfilter to 2.4.0 and 2.4.1+ don't automatically treat
non-required as allowing empty and `false` is treated as empty so making this change now should allow us to
move beyond 2.4.0. I haven't made that change now just in case there are any further issues.