CGIV-8255: Added a support for linked stock locations build up by related product links
Linked stock locations only exist within the api, outside the api they act and behave as normal stock location entities. The resonating for this is that it simplifies all the calling code, as they don't need to worry about what type of stock location they are interacting with. This does require more heavy lifting on the api side but the advantages to the calling code out weigh the extra work on the api.