CGIV-9621: Added barcodeNotApplicable to ProductDetail.
Some channels require a barcode but others do not. When listing to the latter we need to allow the users
to flag that no barcode is required to ge passed validation and also because some of those channels (e.g. eBay)
require us to specifically denote that no barcode is required.