CGIV-10677: fixed an issue with linked products where we sometimes ended up fetching ALL stock.
We can end up in a scenario where either there are no StockLocations for a linked product or,
more likely, only StockLocations in a different Location which get ignored which meant
we ended up filtering the stock table with an empty array of IDs i.e. we fetched ALL stock
which is very bad.