CGIV-11329: changed the query in FindIncorrectlyAllocatedStock back to using replace on SKUs but with the right number of backslashes.
We've tried to fix this many times. The issue is with SKUs with underscores, percents or backslashes in them as they're special chars.
We need to use just the right amount of escaping to be able to compare these SKUs with each other.
For underscores and percents: we need to escape them with a backslash, however, because they're defined within single quotes we need two backslashes.
For actual backslashes: we need to double the escaping.