CGIV-11351: fix to StockLocation updating for linked product to invalidate the eTag cache of the related StockLocations.
When an update of a linked StockLocation ocurrs we updated the related entities too.
There was already code to invalidate the redis cache and the nginx cache for these entities but not the eTag cache.
This lead to a problem where you could update two separate StockLocations with the same underlying stock in quick
succession and they would not conflict, overwriting each others changes.