CGIV-11548: use random base64 as the pathId
The previous method to fetch the next pathId was to run a quite expensive query to find the next available int - not necessarily the maximum one, but the first free sequential int e.g. if pathIds 1,2,4,5 existed, then it would have returned 3. This query caused problems when multiple processes accessed the table at the same time